Cast (size_t) to (uint32_t) when printing.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@287 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/bgp/update_test.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/bgp/update_test.cc,v 1.5 2003/01/29 01:05:32 rizzo Exp $"
+#ident "$XORP: xorp/bgp/update_test.cc,v 1.6 2003/01/29 02:09:37 rizzo Exp $"
 
 #include "bgp_module.h"
 #include "config.h"
@@ -57,8 +57,8 @@ test1(unsigned int as_size)
 	seq1.add_as(AsNum(10));
 
     size_t len;
-    fprintf(stderr, "trying size %d wire_length %d\n",
-	as_size, seq1.wire_length());
+    fprintf(stderr, "trying size %d wire_length %u\n",
+	    as_size, (uint32_t)seq1.wire_length());
     const uint8_t *d = seq1.encode(len, NULL);
     AsSegment *seq2 = new AsSegment(d);
     delete[] d;
